+++ minimal_uhdless_demo/src/tcp_simple.cpp
@@ -112,7 +112,7 @@ size_t tcp_simple::recv(void* buff, size_t size, double timeout) {
     do {
         struct timespec ts_timeout;
         ts_timeout.tv_sec = (time_t) timeout;
-        ts_timeout.tv_nsec = (long) ((timeout - ts_timeout.tv_sec) * 1000000000);
+        ts_timeout.tv_nsec = (long) ((timeout - static_cast<double>(ts_timeout.tv_sec)) * 1000000000);
 
         time_spec_t start = get_system_time();
